requirements.txt
fastapi
uvicorn
python-multipart
pypdf
langchain-google-genai
langchain
faiss-cpu
python-dotenv



